+++ src/geoFire.js
@@ -127,6 +127,15 @@ var GeoFire = function(firebaseRef) {
   /*  PUBLIC METHODS  */
   /********************/
   /**
+   * Returns the Firebase instance used to create this GeoFire instance.
+   *
+   * @return {Firebase} The Firebase instance used to create this GeoFire instance.
+   */
+  this.ref = function() {
+    return _firebaseRef;
+  };
+
+  /**
    * Adds the provided key - location pair to Firebase. Returns an empty promise which is fulfilled when the write is complete.
    *
    * If the provided key already exists in this GeoFire, it will be overwritten with the new location value.
@@ -192,6 +201,10 @@ var GeoFire = function(firebaseRef) {
   /*****************/
   /*  CONSTRUCTOR  */
   /*****************/
+  if (firebaseRef instanceof Firebase === false) {
+    throw new Error("firebaseRef must be an instance of Firebase");
+  }
+
   var _firebaseRef = firebaseRef;
 };